Update patch set 1
Patch Set 1: No score

The arrows are in strings like this:

<string name="right_arrow">»»</string>

It makes no difference to me what char is used. It doesn't matter if they are characters or a drawable because the calculator will never try to validate them. I guess the best thing is to accept this and someone can contribute some nice looking arrows later ;)

Patch-set: 1