Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

A few things that I noted on calculator 3 from market (not sure if is the same code but well i'll throw it anyway)

* UI seems not intuitive, does graph gets automatically generated on function input?

* I got this crash on spanish language, perhaps hardocoded "sin" string?
http://pastebin.com/qAHqavSV

Also a little question: What does the pulldown does, if I pull it I just get a black screen without something to work with.

On other side, patchset has a lot of whitespaces and stuff. I really like the idea of bundling such things on calculator though :)

Patch-set: 1