Update patch set 1
Patch Set 1:

I fixed the localization issues when graphing, the coma issue when comas = decimal place, and tried to be more consistent in formatting (replaced all tabs with spaces, all functions end with a closed parenthesis, a space, and then an open curly brace). The code's on my github.
https://github.com/Xlythe/android_packages_apps_Calculator

Patch-set: 1