Update patch set 1
Patch Set 1:

The matrix add button is a little weird because it's scaling down an image. I can pester my friend for a smaller one tomorrow. As for graphing, pinch-to-zoom is fighting with the ViewPager so you get jerky behavior. Well, I'll let you guys test it out. I'll finish it tomorrow - need sleep.

Patch-set: 1