Update patch set 1
Patch Set 1:

Alright, you can zoom in the graph. The zoom buttons work a lot better than pinch-to-zoom, though (Again, because of the ViewPager). It's on my github.

Panning glitches a LOT. To test it out, change:

renderer.setPanEnabled(false, false);

To:

renderer.setPanEnabled(true);

Or if you want to try both of them out, here's an apk (uninstall the market one - the keys wont match)

https://dl.dropbox.com/u/6408566/android_packages_apps_Calculator.apk

Patch-set: 1