Update patch set 1
Patch Set 1:

Will, I tried the latest app you published, no crash now and graph is showing, thanks for the fix.

A few things then based on the current status:

* Would be cool to include an introduction (trebuchet like) on how to use the app, not everyone will get the how-to-graph, but I agree, a button is not the way it would kill the entire UI.
* Matrix layout is MUCH better now, the button fits better. 
* The pulldown bar for the history should be to disabled if no history is found, because otherwise it just shows an empty black screen.

I also found that on graph the values are over the Y axis, but they are fine on X axis, not a biggie though. Overall app is working great when you learn to use it, looking forward to see this is merged

Patch-set: 1