Update patch set 1
Patch Set 1:

Hey Jesús, I'm going to have to dig into trebuchet for that one. It's not something I know how to do yet. I can't do in an hour or two - it'll take a while. (If anyone else wants to do it, please go right ahead)

On the other hand, disabling the pulldown with no history takes a minute to do, so I'll push that in a bit.

That's a limitation of the graphing library I picked. Nothing I can really do about that.

Translations. These changes add a bunch of new text, none of which is translated at the moment. Anyone who wants the pages in their own language, email me and I can put that in.

Patch-set: 1