Update patch set 5
Patch Set 5:

Wow, thanks! I was dreading going through everything and looking for those small formatting issues. I removed the spaces and made the suggested tweaks. :]

I pushed an update with the following changes
-New animated slider for history
-Disable going back to basic panel in settings
-Rearranged the tablet ui so things are less squished
-Aforementioned tweaks

I still want to play around with the history bar so it's more comfortable. Do things like move it next to DELETE on phones in portrait mode, or maybe make the background a gradient.

Patch-set: 5