Update patch set 5
Patch Set 5:

BTW, the history bar should have better velocity tracking or something? It's very hard to bring it down.

Should be, as in launcher, either required velocity (which can be rather low, since it rarely gets dropped accidentally) or required percentage drop (e.g. if the bar is less than 40% away from the new state and is let go at 0 velocity it should drop to the bottom). 

And just to make it look nicer, it should have an animation when let go of and use sliding rather than scaling effect.

Patch-set: 5