Update patch set 5
Patch Set 5:

Alright, feature locked. I'll fix bugs reported, but I won't be adding anything new. That should slow down the differences between here and the market. It was supposed to be feature locked before I submitted it, but I got a massive amount of feedback when Ricardo posted about it on G+. Oh well, I think it's a better app for it.

Patch-set: 5