Update patch set 1
Patch Set 1:

It's not that complicated. A complex noun is "sammensat navneord" in Danish. In English, complex nouns are almost always divided into 2 words, but that's not the case in Danish. But in some cases, you just can't put 2 words together in Danish, it just looks weird. Like "LEDnotifikation" for instance. That's where you'd put a dash into the word.

Take a look at the word "Function panel". If you had to say it out loud in Danish, wouldn't it be "funktionspanel" instead of "funktion panel"?

But it's your translation, you decide. :) I just give my comments.

Patch-set: 1