Update patch set 1
Patch Set 1:

You are absolutely right regarding function panel but this is not a "normal" function, it is a mathematical function that is why I am not quite sure whether to write it in one or two words, but I'll figure that out? It is not only my translation, it is all the users, so it has to be correctly translated and your opinion is appreciated.

Patch-set: 1