Update patch set 1
Patch Set 1:

I wasn't sure if I should change the underlying libraries. If I can, I changed achartengine to allow the X,Y axis to be bolded and I can add that in too.

Also, please tell me if it compiles. :/ The manifest is different from how I'm used to, and it wouldn't compile on my machine. I assumed there was an extra step and that most of you would know what to do. I just copied the res and src folders over and tried to touch as little as possible.

Patch-set: 1