Update patch set 1
Patch Set 1:

renderer.setZoomButtonsVisible still makes it crash hard due to the lack of the zoom button drawables (that's why it was commented out in the previous merge).

Other than that, it appears to be OK. It should be good to merge after M2 is out and feature submission is unfrozen.

Patch-set: 1