Update patch set 4
Patch Set 4:

There is a bug with logarythms (well, not sure if it is still in there, but it was with the initial patch) that is probably related to the translated buttons. If I try to get either the natural or the base-10 logarythm of any number, it just gives me an error. Is this something to be fixed in code or in the translation?

Patch-set: 4