Update patch set 4
Patch Set 4: (5 inline comments)

Daniel Buga:
I haven't received any complaints about log/ln in the app on the market, so that's probably fixed. If you want to test it for me, the market and this code should be 99% identical.
https://play.google.com/store/apps/details?id=com.android2.calculator3

Markus Hoeber:
I replied to a few of your comments. Still, I rely entirely on people like you for translations, so if you still thing I'm wrong, I'll make the changes.

Patch-set: 4