Update patch set 5
Patch Set 5:

I updated the files to the latest versions (Except for Russian translations, mine are more up to date). However, I did it manually. If there's something I need to do to tell git that I'm up-to-date, could you spell it out for me? I'm bad with git.

Including "product" makes the string too long. That's why they only say cross and dot. It's the same in the English translations.

Patch-set: 5