Update patch set 7
Patch Set 7:

The patch had an error in ColorButton.Java that was easily corrected. It fixes part of the crashing problem that plagues Calculator on MDPI devices. Only the graph screen fails on encore (Nook Color, now with cm10). It force closes when switching to the graph screen. Much better than closing as soon as Calculator is opened which happens without this patch.

Patch-set: 7