Update patch set 12
Patch Set 12:

Looking over it again, it looks like it's just a function name that was changed and a bit more math added. Not as big a change as I thought. If the merge error's gone, this patch should be fine. If not, then I'll try abandoning and resubmitting.

Patch-set: 12