Update patch set 13
Patch Set 13: I would prefer that you didn't submit this

'E' constant is currently broken in decimal mode and interferes with the hexadecimal 'E'. 'E' constant should be in lowercase just like the vanilla calculator.

Also, the history bar is a bad idea as the whole keyboard should be moving down.

One more thing, I think that the matrix operations are pointless in a calculator like this. It's hard to understand how to use it. Thus it should be removed.

Patch-set: 13