Update patch set 14
Patch Set 14:

Thanks Danny, for the coding suggestions. You're right, they make the code cleaner. :) I left in the setBackground / setBackgroundDrawable bit. From ICS to Jellybean, setBackgroundDrawable is deprecated and setBackground is added. I didn't want to make it JB only over such a small change, so that check was added.

Clément, thanks for the info on e, the constant, acting like E the hex number. That was the case in both French and Italian, so I fixed it in both places.

This new history is supposed to go on top of the calculator, instead of pushing everything down like the old one did. I think it's a neat effect. :)

I agree, the matrix panel lacks polish and fluidity. I think it handles basic matrix equations well enough, though, so I'd like to leave it in. Hopefully it'll inspire someone else to come up with a better idea for matrix math on a touch device. Every page can be disabled in settings, so feel free to disable the matrix panel.

Patch-set: 14