Update patch set 14
Patch Set 14:

It's a "feature". Maybe not the must useful of features, but it's supposed to be that way.

Again, I think it works well enough that I'd like people using it. You've got me thinking how I can improve it, though. But I'll save that for another patch.

Patch-set: 14