Update patch set 1
Patch Set 1:

Hmm. It appears that if I translate some of the strings, it would change the interface not in a very good way.

See: http://imgur.com/2kOVh VS http://imgur.com/ctmcb

What do you guys think? Leave it as is like in the screenshot or not translate those 2 strings so that it doesn't look weird?

Patch-set: 1