Update patch set 2
Patch Set 2: Looks good to me, but someone else must approve

I think this is good for now as an initial translation. Further translations would kill UI in a bad way, and I'd rather that not happen. If someone could think of a way around the UI issue, I'd be glad.

Patch-set: 2