Update patch set 1
Patch Set 1:

Hmm, if I mimicked the glow that happens in the actionbar, it would either be a rectangular pillar of glow-ness or a floating square. Neither would look that great. A custom glow effect that follows the edges of the image might look better, but that'll take some time to design so forgive me if I save it for another commit (but I'll keep it in mind).

Patch-set: 1