Update patch set 5
Patch Set 5: Looks good to me, but someone else must approve

Very good update :) 

The only thing what I don`t like is the version bar in the settings of the app. By the way - is it your purpose that there is no version number in the manifest so that the build version with all informations is displayed?

I would like it much better when the version is displayed in an 'About' category like it is in CMFileManager or Trebuchet.

Patch-set: 5