Update patch set 6
Patch Set 6:

Well, this is a long-size/long-time patch and tests are not building automatically, so I will not oppose to submit it in a separate change (but should be fixed). i'll removed my -1 if ux doesn't have any problem on submit it.

nebkat, david, danny, what do you think?

Patch-set: 6