Update patch set 2
Patch Set 2:

It's a nice idea to style the menu button like the navigation bar buttons, but the current implementation really leaves a lot to be desired. It gets abruptly cut off by the number area and isn't animated like the nav buttons.


The standard blue click background in this commit for it is consistent with the rest of the OS and also fits in nicely with the blue effect that shows up when you press the other buttons. For these reasons, in my personal opinion the blue effect would be better, unless someone improves the current one.


(Just sharing my two cents.)

Patch-set: 2