Update patch set 2
Patch Set 2:

Add android:clipChildren="false" to the parent LinearLayout and it'll stop clipping the circle. The animation is harder - I only saw support for Fade (not Shrink), and its buggy when clipChildren is false.

-------

If you want to use a solid color instead, I appear to be in the minority. But fix the translations - remove the translated pref_application Strings (Or edit your language's translation instead).

Patch-set: 2