Update patch set 1
Patch Set 1: Do not submit

Thanks, but crunching all the images in place creates nasty merge conflicts so we won't be merging this. We used to do this with the "squisher" script as a post-build step, but we've retired it now that devices have plenty of storage. The overall savings ends up being only about 7MB so we decided to take a shorter build time.

Patch-set: 1