Update patch set 3
Patch Set 3:

What's the problem with supporting , as decimal separator? As an example, German uses '.' as grouping character and ',' as decimal separator. This doesn't look unimplementable to me?

There's no need to define it in XML, BTW: java.text.DecimalFormatSymbols has all the required characters, and there's also methods to format and parse locale-specific decimal numbers.

IMHO, willlingly not supporting i18n is not a nice move.

Patch-set: 3