Update patch set 3
Patch Set 3:

I manually add the digit grouping because:
a) Its of an equation, not a single number, so I'm iterating through text
b) I couldn't find a way that handles different bases (HEX, DEC).

And commas are difficult because matrices are hard coded to use them as separators in their text representations.

The library that solves the equations requires everything to be in US Locale, so . instead of , and all that stuff. I have to translate it behind the scenes and if a coma is sometimes a decimal separator and sometimes a digit group character, it becomes a mess. If I can create a double given a String and a locale, and then turn it back to a string, this might not be tooooo bad. 

I'm not purposefully trying to harm users. Its just a complicated problem (perhaps because I've made it complicated). Feel free to try your hand at it.

Patch-set: 3
Reviewer: Will Harmon <5633@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0