Update patch set 1
Patch Set 1:

The EditText (that holds the equation) returns a de-comified version of the equation, so comas can be safely ignored. They are purely cosmetic.

The math for changing modes handles decimal points, and calculates the correct value. If this is broken, then I'll fix it, but it's working as far as I know.

Patch-set: 1