Update patch set 1
Patch Set 1:

>>The check the https://jira.cyanogenmod.org/browse/CYAN-1817 issue.

I don't get that issue, on 4.2 (stable) or 4.3 (nightly from last week). I do get 0.9999. Weird.

>>Actually, 1.2 in dec, returns 1.33333333 in hex. AFAIK, bin and hex shouldn't have floating number, so what is the value of 1.2 in hex?

Bin and hex have floating numbers, they're just not used often. It's the same concept as in decimal. The reason 1.2 in dec is 1.3333 in hex can be thought of like this... 0.2 is 20% of dec. hex has 16 characters, so find 20% of 16 and you get 0.3333. Then add the 1 to get 1.3333. (If we used larger numbers, we'd have to account for ABCDEF so we're lucky with this example)

>>BTW, negative numbers conversions doens't work: -12 should be FFFFFFFFFFFFFFF4, not -C

It uses proper negatives. You dont say -12 is actually 999999999995. We have an infinite number of characters (unlike a computer which limits itself to blocks of 8 and 16).

Patch-set: 1