Update patch set 1
Patch Set 1:

Fixed - I'll push a release soon. Working on a new feature so my code's a bit messed up.

If anyone needs it NOW (for some reason), add:

if(split[1].length() > 13) {
    split[1] = split[1].substring(0, 13);
}

to newBase(String, int, int) in BaseModule. It's a decimal, and we're rounding anyway, so I just catch the error.

Patch-set: 1