Update patch set 3
Patch Set 3:

Will, we found out that changing the Separators for dec, bin, and hex and decimal point as we use them in Dutch do not work. It always gives an ERROR in the display now, even on the most basic calculations.

Can you explain what is wrong? In Dutch a , is used as a seperator for decimals (for example 2,5) and a . is used for numbers in the thousands (for example 100.000 = hundredthousand)

This is how it was in Dutch (now reverted/removed)


<string name="dec_separator">.</string>		

<string name="bin_separator">\u0020</string>	
	
<string name="hex_separator">\u0020</string>	
	
<string name="dot">,</string>		

<string name="matrix_separator">.</string>


Any idea?
Thanks

Patch-set: 3
Reviewer: Eddy Witkamp <2881@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0