Update patch set 2
Patch Set 2:

ROMs with the theme engine could theme the dark theme however they'd like. There's already a bit of duplication just in having a light theme. Well, I'm indifferent on the theme and the original designer wishes the theme never existed. We'd both be fine with whatever direction the community wants to take the light theme.

Patch-set: 2