Update patch set 2
Patch Set 2: Code-Review-1

Not a fan. With the addition of graphing and other capabilities, this is less a 'stock' app than it is a 'CM' app. 

Not only does this fit KK's guidelines, but I like it (was just working on a change to set light theme by default).

Patch-set: 2
Label: Code-Review=-1