Update patch set 2
Patch Set 2:

Strange. Are you using proguard to obfuscate the classnames? And if so, is it possible to turn that off?

I'm using reflection to look up R.color, R.drawable, R.string, etc. So if it's no longer R.color but R.asdfasfa, it crashes. If you can't turn off proguard, I can rewrite it, but it won't be as clean.

Patch-set: 2