Update patch set 8
Patch Set 8:

I do ask that the Theme Engine stay in. It's a way for me (and the designers) to profit off the app by providing a fringe feature. We do not work for CyanogenMod and receive no benefits for our efforts otherwise.

Other reasons. Calculator exists in many other ROMs, as well as on the Play Store, and there is no desire to maintain multiple branches of the app. The Theme Engine lets us play with different ideas for the app, one of which is a hopeful "Stock CyanogenMod" theme matching the white/light blue color scheme of newer apps. And lastly, Calculator has a light theme already that, without this engine, would be gone.

We've been building/maintaining this app for almost 2 years. Give us a little leeway? :]

Patch-set: 8