Update patch set 8
Patch Set 8: Code-Review-1

I have to agree with the general tone of Clark's discussion.

My concerns:
1) As stated, we have a highly capable theme engine; these one-off theme's choices should be removed from here and apps like File Manager. Having duplicate ways and concepts of 'themes' introduces unnecessary complexity. 

2) The store links to a json file on a remote server not in control by CM or reviewed by Gerrit. If this functionality is to get merged, I would want this to be hosted on a CM controlled server with updates possible to the json via Gerrit. This not only allows for transparency and version control of the listings, but allows for others to build 'themes' and get featured in this list. 

As it stands, for understandable reasons, only your own themes are available. This doesn't make the list seem very 'open'. There is also the potential for user confusion in the current implementation, purchasers may assume content available in this store is to benefit the CM project as a whole, not an individual market developer (though you are indeed both - just making the crude distinction). 

So the questions I have would be: Do we really want a second theming implementation, and if so, shouldn't the listing be open?

Patch-set: 8