Update patch set 8
Patch Set 8:

I'm fine with the store being controlled by CyanogenMod. It's an understandable request.

Last night, I wrote a short guide on creating themes. It's not yet pushed, but there's a ? button in the store detailing themes and a (simplistic) submission process. The idea is that anyone can create/share a theme.

http://xlythe.com/calculator/theme.html
https://www.dropbox.com/s/dqxttw73wgiutdh/Screenshot%202014-04-21%2021.34.39.png
https://www.dropbox.com/s/f0yfabd1rlspezn/Screenshot%202014-04-21%2021.34.33.png

The road map is to eventually also allow Extensions, additional pages that can be added to the calculator. That's still being worked on, since my current method doesn't work with encrypted apps (and all paid apps on the Play Store are encrypted).

Patch-set: 8