Update patch set 10
Patch Set 10: Code-Review+1

To immediately solve the issue of the store being controlled by me, I have added a txt file (themes.json) to the root of Calculator and pointed the Store to look for that file on GitHub. That means the Store will be broken until this patch is commited. Edit it to http://xlythe.com/calculator/store/themes.json to see it working.

A txt file is a bit rudimentary but it works. Themes can be submitted to the store via patches, making it controlled by CyanogenMod. That means major blockers for this patch should be gone. Themes can migrate to CyanogenMod's engine when there's support to just theme a single app.

Patch-set: 10