Update patch set 14
Patch Set 14:

Some things I noticed:

The graph view (zoom and offset) is reset on rotation.

Using pinch to zoom in the graph can result in unreadable axis labels (http://i.imgur.com/nsoxrdp.png).

Does it make sense to nest matrices (http://i.imgur.com/XDzbfiD.png)? The only use-case I can think of is if you want the determinant of the inner matrix. But that's not possible to do because the closing bracket for "det(" is added as the last character resulting in an invalid formula like: "[[ det([[1 2] [3 4]] 5] [6 7]]). Adding rows and columns to the inner matrix is also not possible. After rotation this happens: http://i.imgur.com/trJ3qUH.png.

I think the "det" button should be a toggle button to insert "det(" and ")" so I can write "det([[1 2] [3 4]]) * 2" instead of having an invalid formula.

I often cannot see if I moved the cursor before, after or between matrices.

I get a black screen and the following logcat messages http://pastebin.com/raw.php?i=hiMv8Ste when clicking on "Store". I understand why it doesn't work yet but there should be at least a message that the theme store is currently unavailable. Also check if a network connection is available in the first place.

Patch-set: 14