Update patch set 15
Patch Set 15: Code-Review+1

Forgot my tweak to the Store - It now says to submit themes to review.cyanogenmod.org again, instead of to me.

The app is managed on the Play Store for a week or two before being submitted here. Bugs are pretty much always caught. The biggest issues then are build problems, followed by formatting. Ricardo has, historically, given me some leeway on formatting... Breaking it into smaller patches is an even bigger headache - everything is intertwined - so it increases the risk of a bug and makes it hard to provide fixes because everything has to be ported to the patch. At least as one patch, I can just run a formatter (theoretically. In practice, apparently not).

Thomas Wendt - You're right. You shouldn't be able to make a matrix inside a matrix. That bug is probably because we turned off text verification to allow sin,cos,etc inside the matrix. I doubt a fix will be in this patch, though. The graph resetting itself on rotation is noted too.

Patch-set: 15