Update patch set 18
Patch Set 18:

Normally the should be a copyright header in every xml and java file so I added one to the files where they were missing. I didn't removed the copyright in the files which come from AOSP.  You can compare it with AOSP... every files has it's copyright but also a CM copyright because we also made changes to those files.

The whitespace changes are just for readability. And they are also just in files which doesn't came from AOSP so I think that should be no problem.

Patch-set: 18