Update patch set 22
Patch Set 22:

Strange. The version on the Play Store works fine - no crashes. Going by the error, it looks like a style="@..." tag accidentally got lost on one of the layout-land files.

I can try pulling down the code tonight (or more likely this weekend). Is it only crashing when rotating, or can the app not even open in landscape?

Patch-set: 22