Update patch set 32
Patch Set 32: Code-Review-1

*Opens can of worms*

This will not be merged with the Store component. I recognize this is a shitty conclusion to my earlier discussion on patchset 8, but introducing multiple definitions of 'Themes' outside of the global theme engine itself is something we need to get away from. Under this same idealogy, I will be moving to remove other 'themes' elements from apps (LockClock, SMS, etc). 

Now, that said, if per app theming or some other advanced mode is required, this can be something worked on for the ThemeChooser itself. 

Where this leaves this patchset (and to a larger degree Will's support of this now-fork) is where this discussion should move to; but the Store/Theme element is a blocker.

Patch-set: 32