Update patch set 32
Patch Set 32:

The CM Theme Engine doesn't handle per-app theming well, and has no support for theming fonts of specific apps. If (or when) those features are added, I agree, one engine to rule them all. However, after asking on Google+, it doesn't seem like those are considered pressing requests.

It's 2 lines to hide the Store and theme-switcher preference in the Calculator. And I've already updated all the themes on the store to support both the CM Theme Engine and the app's native theme engine. Why not accept this and disable the native theme support when ready? The current light theme in the calculator will be much harder to remove than the 2 lines in this patch. And you have my support, as the original dev, when the time comes.

Patch-set: 32