Update patch set 32
Patch Set 32:

I can't speak to timelines at this moment, but your assessment on the fonts item is accurate.  This doesn't mean its ruled out permanently though. 

If it is just 2 lines to hide the store, then honestly, we should just hide the store. Introducing the store/theme element now, and then backing it out in 2/3/4 months would just further introduce confusion to the end user. 

I honestly appreciate your continued support of this application for CM and I recognize the need for consistency in maintaining two codebases - but I believe this decision needs to be made.

Patch-set: 32