Update patch set 32
Patch Set 32:

CM is planning a store of its own. Can we add a feature request to add tags (ie: calculator) and an API to fetch all themes with the tag? The native store can then switch over, same feature set but using the official store for data. Or is that not desired?

I do have a vested interest in getting my themes noticed because most of them are paid. Its selfish of me, but I hope its understandable... Extensions are also on the roadmap, which would similarly be installed with a separate apk, and they're planned as a tab on the store page.

Patch-set: 32