Update patch set 32
Patch Set 32:

For clarity, the 'ThemeShowcase' for themes is a Cyanogen Inc venture, not something that will be introduced into CM mainline or shipped in nightlies. 

The goal there is to have it available to download via the Play Store or w/e, but not 'bloat' the core CM experience. Those that want the extra app can get it. 

Adding tags or expanding the search API's for the showcase app is feasible. 

No need to explain your interests, they are understandable and really fairly reasonable - however, as seen with the Showcase component, I am of the mindset that the core CM experience should not be the vehicle for this. We walked away from Rom Manager and even ADW under the same pretenses.

Patch-set: 32