Update patch set 32
Patch Set 32:

Hmm, the Store only shows up when it detects the Play Store as installed. A similar check could be added for the ThemeShowcase.

If I disabled the Store, but left the current theme engine in (with the intention to disable it when there's better support for my use case in the CM Engine), could we get this merged today?

Patch-set: 32