Update patch set 33
Patch Set 33:

Something is VERY WRONG with the calculator since this patch.

To me it seems to be a problem with the decimal point in other languages, like Dutch. It uses a dot as a decimal separator, not a comma. This seems to have been removed in this patch.


This was brought to my attention in this post: 

https://plus.google.com/111355926593696741923/posts/LLUtoYFvHTs



Problem 1:

Type 9000 and it becomes 9,000. In Dutch that is the same as 9,0. It should be 9.000 (with a dot, not a comma)


Problem 2: 

unable to use the decimal point. When I tap it, nothing happens. Tapping 9,2 results in 92


Problem 3: 

calculation is totally wrong. For example
9321 / 2 = 46605    (wth)



Please take a look at the decimal problem.

Patch-set: 33
Reviewer: Eddy Witkamp <2881@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0