Update patch set 33
Patch Set 33:

On second thought, this could also be a translation problem of the dec_separator string.

Changing it to a Unicode character has triggered Crowdin to make this string untranslated. Most translators have just copied the Unicode in their translation.


Is there a way to make the seperator untranslatable and dependent on the language that is chosen??

Patch-set: 33