Update patch set 2
Patch Set 2:

Oh, yea, true. Yours was ignoring the "enabled" flag. That's why I changed it originally; the &= was so it obeyed both flags. I ended up removing the flag because seeing 7 and 4 fade black when you click bin is nice. But I don't like being able to click 7, 4, 1, and . when the panel is open...

Anyway, the &= flag is probably better in case someone modifies enabled in the future. In my opinion.

Patch-set: 2