Update patch set 1
Patch Set 1:

Expressions were normalized but weren't localized after they were solved.
Imagine:
',' as decimal point
'.' as decimal separator

',' was normalized to '.' but was never localized again.
TextChangedListener in CalculatorEditText took '.' as decimal separator and removed it.

The order of the replacements in CalculatorExpressionTokenizer had to be changed because '.' would have been localized to ',' and after that it would have been localized to the matrix separator.

PS: Please excuse my english.

Patch-set: 1