Update patch set 1
Patch Set 1:

(1 comment)

I'd recommend fixing the background of the first child of mView.  If the view fits without adjusting the Y value then set it to floating_calc_background otherwise set it to a resource ID of 0 to clear it out.

The little up caret doesn't make sense when the fab is to the right of the floating calculator view.

Patch-set: 1